Correction, création ennemy bien gérer dans la classe correspondante
